% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{cellTypeClassifier}
\alias{cellTypeClassifier}
\title{Train and Predict Cell Type Classifier Using caret}
\usage{
cellTypeClassifier(trainData, testData, model = "lda")
}
\arguments{
\item{trainData}{Data frame; training data with cell type labels.}

\item{testData}{Data frame; test data to predict cell types.}

\item{model}{Character; machine learning model to use, default is "lda".}
}
\value{
Vector of predicted cell types for the test data.
}
\description{
Trains a cell type classifier using LDA (or other specified models) and performs
predictions on test data.
}
