% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMD_Vae.R
\name{computeMMD}
\alias{computeMMD}
\title{Compute Maximum Mean Discrepancy (MMD) Loss}
\usage{
computeMMD(x, y, sigmaSqr = 1)
}
\arguments{
\item{x}{Tensor; first input distribution.}

\item{y}{Tensor; second input distribution.}

\item{sigmaSqr}{Numeric; variance parameter for RBF kernel, default is 1.0.}
}
\value{
Tensor; scalar MMD loss.
}
\description{
Calculates the MMD loss between two distributions using an RBF kernel.
}
