% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMD_Vae.R
\name{decodeSWAE}
\alias{decodeSWAE}
\title{Decode Latent Representations with Enhanced SWAE}
\usage{
decodeSWAE(model, latentData)
}
\arguments{
\item{model}{Trained enhanced SWAE model from trainSWAEEnhanced.}

\item{latentData}{Matrix of latent representations.}
}
\value{
Matrix of reconstructed data.
}
\description{
Reconstruct data from latent representations using trained enhanced SWAE decoder.
}
