% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMD_Vae.R
\name{encodeSWAE}
\alias{encodeSWAE}
\title{Encode Data with Enhanced SWAE}
\usage{
encodeSWAE(model, newData)
}
\arguments{
\item{model}{Trained enhanced SWAE model from trainSWAEEnhanced.}

\item{newData}{Matrix or data frame of data to encode.}
}
\value{
Matrix of latent representations (cells x latent dimensions).
}
\description{
Get latent representations from a trained enhanced SWAE model.
}
