% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{findChildren}
\alias{findChildren}
\title{findChildren}
\usage{
findChildren(tree)
}
\arguments{
\item{tree}{a ggtree object}
}
\value{
a ggtree object with the data containing a column with the clusters
contained in each node
}
\description{
findChildren
}
\examples{
# Load sample data
data(sample_tree)

# Apply findChildren to tree
tree_with_children <- findChildren(sample_tree$tree)

# Check clusters
print(tree_with_children$data$clusters)
}
