% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{generateGroups}
\alias{generateGroups}
\title{Generate Hierarchical Groups from Clustering Tree}
\usage{
generateGroups(tree, nClust = 20, proportions, means, type = "all")
}
\arguments{
\item{tree}{A hierarchical clustering tree generated from \code{generateTree}.}

\item{proportions}{Data frame of cell type proportions.}

\item{means}{Data frame of marker means.}

\item{type}{Character; group type to generate ("all", "prop", or "mean").}

\item{nclust}{Integer; number of clusters to generate.}
}
\value{
List of variable groups for overlapping group lasso.
}
\description{
Creates variable groups for overlapping group lasso based on hierarchical
clustering tree structure, optionally using proportions or means.
}
\examples{
# Load sample data
data(sample_data_logit)
data(sample_markerMeans)
data(sample_tree)

# Generate groups for overlapping group lasso
groups <- generateGroups(
  tree = sample_tree$tree,
  nClust = 5,
  proportions = sample_data_logit,
  means = sample_markerMeans,
  type = "all"
)

length(groups)
sapply(groups[1:5], length)
}
