% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{generateTree}
\alias{generateTree}
\title{Generate Hierarchical Clustering Tree for Group Lasso}
\usage{
generateTree(features, method = "ward")
}
\arguments{
\item{features}{Numeric matrix of feature data.}

\item{method}{Character; clustering method, default is "ward".}
}
\value{
List containing the hierarchical tree structure and node order.
}
\description{
Constructs a hierarchical clustering tree from feature correlations for use
in group lasso. Correlations are converted to distances, and the tree is
created using the specified clustering method.
}
