% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{Plot Heatmap of Marker Means Coefficients from Group Lasso Model}
\usage{
plotHeatmap(fit, type = "cluster", order = NULL, markerOrder = NULL)
}
\arguments{
\item{fit}{Fitted group lasso model.}

\item{type}{Character; type of clustering, default is "cluster".}

\item{order}{Optional vector to order clusters in the heatmap.}

\item{markerOrder}{Optional vector to order markers in the heatmap.}
}
\value{
ggplot object representing the heatmap.
}
\description{
Creates a heatmap of coefficients from the group lasso model, grouped by cell type clusters.
}
