% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{plotOverlapTree}
\alias{plotOverlapTree}
\title{Plot Overlap Group Lasso Feature Tree}
\usage{
plotOverlapTree(modelFit, trainingData, treeData, nodesToRemove)
}
\arguments{
\item{modelFit}{Fitted group lasso model.}

\item{trainingData}{Data frame used for training the model.}

\item{treeData}{Data containing hierarchical structure of features.}

\item{nodesToRemove}{Vector of nodes to remove from the hierarchy.}
}
\value{
A ggplot object representing the feature tree.
}
\description{
Visualizes the feature hierarchy from a group lasso model using a dendrogram.
Highlights features based on their beta coefficients, with different colors
representing the internal and leaf nodes.
}
