% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{plotSigFeatures}
\alias{plotSigFeatures}
\title{Plot Expression of Significant Features}
\usage{
plotSigFeatures(stats, sigFeatures, outcome, title = "", type = "density")
}
\arguments{
\item{stats}{Data frame containing statistical results for each feature.}

\item{sigFeatures}{Data frame of significant features and their expression levels.}

\item{outcome}{Character; name of the outcome variable to group by.}

\item{title}{Character; title of the plot, default is an empty string.}

\item{test}{Logical; if TRUE, applies significance testing and displays p-values, default is FALSE.}
}
\value{
A ggplot object representing the boxplot of significant features.
}
\description{
Creates a boxplot to visualize the expression levels of significant features
identified by the model, grouped by outcome. Optionally overlays statistical
significance labels on the plot.
}
