% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_groups}
\alias{sample_groups}
\title{Sample Group Structure for Overlapping Group Lasso}
\format{
A list of 10 numeric vectors, where each vector contains
variable indices (1-30) that form a group. Variables can appear in
multiple groups (overlapping).
}
\source{
Manually defined based on biological relationships
}
\usage{
sample_groups
}
\description{
A list defining overlapping groups of variables for the group lasso penalty.
Each element contains indices of variables that belong to a group.
}
\details{
Groups 1-3 contain proportion features (indices 1-5),
groups 4-8 contain cell-type-specific marker means,
and groups 9-10 contain marker-specific means across cell types.
}
\examples{
data(sample_groups)
length(sample_groups)
sapply(sample_groups, length)
}
\keyword{datasets}
