% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_markers}
\alias{sample_markers}
\title{Sample CyTOF Markers}
\format{
A character vector of length 5 containing:
\describe{
\item{CD3}{T cell marker}
\item{CD4}{Helper T cell marker}
\item{CD8}{Cytotoxic T cell marker}
\item{CD19}{B cell marker}
\item{CD14}{Monocyte marker}
}
}
\source{
Synthetic data
}
\usage{
sample_markers
}
\description{
A character vector containing the names of 5 key CyTOF markers used
in the example datasets.
}
\examples{
data(sample_markers)
print(sample_markers)
}
\keyword{datasets}
