% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_tree}
\alias{sample_tree}
\title{Sample Hierarchical Tree Structure}
\format{
A list containing:
\describe{
\item{tree}{A list with a data component containing tree structure}
\item{tree$data}{Data frame with node information including clusters}
\item{order}{Numeric vector with hierarchical clustering order}
}
}
\source{
Generated using hierarchical clustering on sample_data_logit
}
\usage{
sample_tree
}
\description{
A hierarchical tree structure generated from clustering cell type proportions.
Used for defining overlapping groups in group lasso.
}
\examples{
data(sample_tree)
names(sample_tree)
nrow(sample_tree$tree$data)
}
\keyword{datasets}
