% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{selectAlpha}
\alias{selectAlpha}
\title{Select Optimal Alpha for Group Lasso Model}
\usage{
selectAlpha(
  xTrain,
  yTrain,
  groups,
  penalty = "cMCP",
  weights,
  alphaSearch = seq(0.01, 0.1, by = 0.01),
  seed = 1994
)
}
\arguments{
\item{xTrain}{Matrix of training features.}

\item{yTrain}{Vector of training labels.}

\item{groups}{List of groups for overlapping group lasso.}

\item{penalty}{Character; penalty type, default is "cMCP".}

\item{weights}{Vector of weights for the lasso penalty.}

\item{alphaSearch}{Numeric vector of alpha values to search.}

\item{seed}{Integer; random seed for reproducibility.}
}
\value{
A list with the best fitted model, best BIC, best alpha, and BIC values for each alpha.
}
\description{
Searches for the best alpha value for the group lasso model by minimizing BIC.
}
