% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{plotElbow}
\alias{plotElbow}
\title{Plot Elbow Graph for BICs}
\usage{
plotElbow(bics)
}
\arguments{
\item{bics}{Data frame with BIC values for each alpha.}
}
\value{
A ggplot object showing the elbow plot.
}
\description{
Generates an elbow plot of BIC values to aid in selecting the optimal alpha value.
}
