% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling_and_visualization.R
\name{visualiseModelTree}
\alias{visualiseModelTree}
\title{Visualize Group Lasso Model Tree}
\usage{
visualiseModelTree(
  fit,
  tree,
  type = "cluster",
  heatmap = TRUE,
  trainingData,
  nodesToRemove = NULL,
  title = "Feature Tree"
)
}
\arguments{
\item{fit}{Fitted group lasso model.}

\item{tree}{Tree structure data used to define feature hierarchies.}

\item{type}{Character; specifies type of grouping, either "cluster" or "marker", default is "cluster".}

\item{heatmap}{Logical; if TRUE, overlays a heatmap of feature coefficients, default is TRUE.}

\item{trainingData}{Data frame of training data for use in the visualization.}

\item{nodesToRemove}{Vector of nodes to remove from the tree.}

\item{title}{Character; title of the plot, default is "Feature Tree".}
}
\value{
A ggplot object representing the dendrogram with optional heatmap overlay.
}
\description{
Creates a dendrogram visualization for the group lasso model. Optionally,
displays a heatmap overlay of the feature coefficients.
}
